refactor: regenerate migrations

+++ Commerce/Data/Entities/VariantAttribute.cs
@@ -3,7 +3,7 @@ namespace Commerce.Data.Entites;
 public class VariantAttribute
 {
     public ProductVariant Variant { get; set; } = null!;
-    public int VariantId { get; set; }
+    public long VariantId { get; set; }
     public string Name { get; set; } = null!;
     public string Value { get; set; } = null!;
 }
